pyproject.toml
[project]
name = "django-drf"
version = "0.1.0"
description = ""
authors = [
    {name = "Henhe",email = "Henhe@Rambler.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-camel-case (>=1.4.2,<2.0.0)",
    "django-seed (>=0.3.1,<0.4.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "celery (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
